Remove ServerResolver implementation for Qt 4
Qt 4 used a different implementation because QDnsLookup was introduced in Qt 5.

Unlike the complete implementation, the Qt 4 compatible one only supported simple lookups (e.g. hostname -> IP address).
